せどり用/requirements.txt
click>=8.1.0
tabulate>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.28.0
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
